Slightly rework conversion code in count64generic
While this change slows the code down on arm64, it's an
improvement on 32 bit targets.  Here are results for 386:

Count64/generic/1000-8        465MB/s ± 0%   494MB/s ± 0%  +6.22%  (p=0.000 n=9+8)
Count64/generic/10000-8      1.46GB/s ± 1%  1.50GB/s ± 1%  +3.09%  (p=0.000 n=9+10)
Count64/generic/100000-8     1.86GB/s ± 1%  1.89GB/s ± 1%  +1.82%  (p=0.000 n=9+9)
Count64/generic/1000000-8    1.91GB/s ± 1%  1.93GB/s ± 1%  +1.44%  (p=0.000 n=9+10)
Count64/generic/10000000-8   1.89GB/s ± 1%  1.92GB/s ± 1%  +1.52%  (p=0.000 n=8+9)
Count64/generic/100000000-8  1.89GB/s ± 1%  1.91GB/s ± 2%  +0.95%  (p=0.013 n=9+10)

+++ generic.go
@@ -250,14 +250,14 @@ func count64generic(counts *[64]int, buf []uint64) {
 		dcba3 := ba1>>2&0x3333333333333333 | dc1&0xcccccccccccccccc
 
 		// split counters for better performance on 32 bit systems
-		dcba0l := uint32(dcba0)
-		dcba0h := uint32(dcba0 >> 32)
-		dcba1l := uint32(dcba1)
-		dcba1h := uint32(dcba1 >> 32)
-		dcba2l := uint32(dcba2)
-		dcba2h := uint32(dcba2 >> 32)
-		dcba3l := uint32(dcba3)
-		dcba3h := uint32(dcba3 >> 32)
+		dcba0l := uint(uint32(dcba0))
+		dcba0h := uint(dcba0 >> 32)
+		dcba1l := uint(uint32(dcba1))
+		dcba1h := uint(dcba1 >> 32)
+		dcba2l := uint(uint32(dcba2))
+		dcba2h := uint(dcba2 >> 32)
+		dcba3l := uint(uint32(dcba3))
+		dcba3h := uint(dcba3 >> 32)
 
 		// add to counters
 		counts[0] += int(dcba0l & 0x0f)
